requirements.txt
argparse
cryptography
python-magic-bin
tgcrypto